add detection test file
